requirements.txt
alembic==0.9.6
astroid==1.6.0
attrs==17.3.0
beautifulsoup4==4.6.0
coverage==4.4.2
decorator==4.1.2
-e git+git@github.com:ricardosiri68/dingen-pyramid.git@9605ea8fc6d0c5c7c1a88ba0d8e52c58e117e036#egg=dingen_pyramid
factory-boy==2.9.2
Faker==0.8.10
flake8==3.5.0
hupper==1.0
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.11.1
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
parso==0.1.1
PasteDeploy==1.5.2
pexpect==4.3.1
pickleshare==0.7.4
pkg-resources==0.0.0
plaster==1.0
plaster-pastedeploy==0.4.2
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.1
pyramid==1.9.1
pyramid-debugtoolbar==4.3
pyramid-jinja2==2.7
pyramid-mako==1.0.2
pyramid-retry==0.5
pyramid-tm==2.2
pytest==3.3.0
pytest-cov==2.5.1
python-dateutil==2.6.1
python-editor==1.0.3
repoze.lru==0.7
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.15
text-unidecode==1.1
traitlets==4.3.2
transaction==2.1.2
translationstring==1.3
venusian==1.1.0
waitress==1.1.0
wcwidth==0.1.7
WebOb==1.7.4
WebTest==2.0.29
wrapt==1.10.11
zope.deprecation==4.3.0
zope.interface==4.4.3
zope.sqlalchemy==0.7.7
